+++ src/suou/sqlalchemy/__init__.py
@@ -157,13 +157,17 @@ def require_auth_base(cls: type[DeclarativeBase], *, src: AuthSrc, column: str |
 
 
 from .asyncio import SQLAlchemy, AsyncSelectPagination, async_query
-from .orm import id_column, snowflake_column, match_column, match_constraint, bool_column, declarative_base, author_pair, age_pair, bound_fk, unbound_fk, want_column
+from .orm import (
+    id_column, snowflake_column, match_column, match_constraint, bool_column, declarative_base, 
+    author_pair, age_pair, bound_fk, unbound_fk, want_column, a_relationship, BitSelector, secret_column
+)
 
 # Optional dependency: do not import into __init__.py
 __all__ = (
     'IdType', 'id_column', 'snowflake_column', 'entity_base', 'declarative_base', 'token_signer',
     'match_column', 'match_constraint', 'bool_column', 'parent_children',
     'author_pair', 'age_pair', 'bound_fk', 'unbound_fk', 'want_column',
+    'a_relationship', 'BitSelector', 'secret_column',
     # .asyncio
     'SQLAlchemy', 'AsyncSelectPagination', 'async_query'
 )